Add Netflix-style welcome and restructure pages
